chore(routing): reorganise code
