+++ demo/routingV2/src/debug.mjs
@@ -1,0 +1,110 @@
+import * as joint from '../../../joint.mjs';
+
+export const debugConf = {
+    showGraph: false,
+    showGrid: false,
+    showGetInBox: false,
+    gridBenchmark: false,
+    routerBenchmark: true,
+}
+export const debugStore = {
+    gridPrinted: false,
+    graphPrinted: false,
+}
+export const debugLog = function () {};
+
+// ======= Visual debugging
+export function showDebugGraph(pathfinder) {
+    const { planner, _step: step, _graph: graph } = pathfinder;
+    const c = new joint.shapes.standard.Circle({
+        type: 'dc',
+        position: { x: 0, y: 0 },
+        size: { width: 10, height: 10 },
+        attrs: {
+            body: {
+                refCx: 0,
+                refCy: 0,
+                pointerEvents: 'none',
+                fill: 'white',
+                stroke: 'red',
+                strokeWidth: 3
+            }
+        }
+    });
+
+    const l = new joint.shapes.standard.Link({
+        type: 'dl',
+        router: {
+            name: 'normal'
+        },
+        attrs: {
+            line: {
+                sourceMarker: null,
+                targetMarker: null,
+                stroke: 'blue',
+                strokeDasharray: '5 5',
+                strokeOpacity: 0.5,
+                strokeWidth: 1
+            }
+        },
+        markup: [{
+            tagName: 'path',
+            selector: 'line',
+            attributes: {
+                'fill': 'none',
+                'pointer-events': 'none'
+            }
+        }],
+        z: -1
+    });
+
+    const debugCells = [];
+    const debugLinks = [];
+
+    planner.graph.verts.forEach(vert => {
+        const ds = c.clone();
+        ds.position(vert.x * step, vert.y * step);
+
+        vert.edges.forEach(edge => {
+            const ls = l.clone();
+            l.source({ x: vert.x * step, y: vert.y * step });
+            l.target({ x: edge.x * step, y: edge.y * step });
+            debugLinks.push(ls);
+        });
+
+        debugCells.push(ds);
+    });
+
+    graph.addCells(debugLinks).addCells(debugCells);
+}
+
+export function showDebugGrid(pathfinder) {
+    const { grid, _step: step, _graph: graph } = pathfinder;
+    const ro = new joint.shapes.standard.Rectangle({
+        type: 'dc',
+        position: { x: 0, y: 0 },
+        size: { width: step, height: step },
+        attrs: {
+            body: {
+                pointerEvents: 'none',
+                fill: '#ff0000',
+                fillOpacity: 0.2,
+                stroke: 'white',
+                strokeWidth: 1
+            }
+        }
+    });
+
+    const gridCells = [];
+    for (let i = 0; i < grid.shape[0]; i++) {
+        for (let j = 0; j < grid.shape[1]; j++) {
+            if (grid.get(i, j) !== 0) {
+                const dc = ro.clone();
+                dc.position(i * step, j * step);
+                gridCells.push(dc);
+            }
+        }
+    }
+
+    graph.addCells(gridCells);
+}
