+++ demo/routingV2/src/models/Obstacle.mjs
@@ -1,0 +1,5 @@
+export default function Obstacle(id, element, grid) {
+    this.index = id;
+    this.cell = element;
+    this.grid = grid;
+}
